pyproject.toml
[tool.poetry]
name = "chitose"
version = "0.1.0"
description = "The backend API for Akari"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
python-dotenv = "^0.21.1"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
better-ipc = "^2.0.1"
pydantic = "^1.10.4"
orjson = "^3.8.5"
prisma = "^0.8.0"
discord-py = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
